better error handling all round
