+++ js/base/errors.js
@@ -17,7 +17,7 @@ function subclass (BaseClass, classes, namespace = {}) {
 
             [className]: class extends BaseClass {
 
-                constructor (message, exchangeId = undefined, httpCode = undefined, httpStatusText = undefined, url = undefined, httpMethod = undefined, responseHeaders = undefined, responseBody = undefined, responseJson = undefined) {
+                constructor (message, exchangeId = undefined,  httpMethod = undefined, httpCode = undefined, httpStatusText = undefined, url = undefined, responseHeaders = undefined, responseBody = undefined, responseJson = undefined) {
                     super (message)
 
                 /*  A workaround to make `instanceof` work on custom Error classes in transpiled ES5.
@@ -31,10 +31,10 @@ function subclass (BaseClass, classes, namespace = {}) {
                     this.message = message
 
                     this.exchangeId = exchangeId
+                    this.httpMethod = httpMethod
                     this.httpCode = httpCode
                     this.httpStatusText = httpStatusText
                     this.url = url
-                    this.httpMethod = httpMethod
                     this.responseHeaders = responseHeaders
                     this.responseBody = responseBody
                     this.responseJson = responseJson
