+++ js/base/functions/string.js
@@ -11,8 +11,9 @@ module.exports =
 
       // hasFetchOHLCV → has_fetch_ohlcv; parseHTTPResponse → parse_http_response
     , unCamelCase: s => s.match (/^[A-Z0-9_]+$/) ? s : (s.replace (/[a-z0-9][A-Z]/g, x => x[0] + '_' + x[1]).replace(/[A-Z0-9][A-Z0-9][a-z]/g, x => x[0] + '_' + x[1] + x[2]).toLowerCase ())
+    , camelCase: s => s.replace (/_(\w)/g, (x, y) => y.toUpperCase ())
 
-    , capitalize: s => s.length
+        , capitalize: s => s.length
                             ? (s.charAt (0).toUpperCase () + s.slice (1))
                             : s
     }
